requirements.txt
aiohttp                   3.8.6
aiosignal                 1.3.1
altgraph                  0.17.4
annotated-types           0.6.0
anyio                     3.7.1
async-timeout             4.0.3
attrs                     23.1.0
beautifulsoup4            4.12.2
certifi                   2023.7.22
chardet                   3.0.4
charset-normalizer        3.3.2
colorspacious             1.1.2
contourpy                 1.2.0
cycler                    0.12.1
deep-translator           1.11.4
exceptiongroup            1.1.3
fonttools                 4.44.0
frozenlist                1.4.0
googletrans               3.0.0
h11                       0.12.0
h2                        3.2.0
hpack                     3.0.0
hstspreload               2023.1.1
httpcore                  0.13.7
httpx                     1.0.0b0
hyperframe                5.2.0
idna                      2.10
kiwisolver                1.4.5
langdetect                1.0.9
mpmath                    1.3.0
multidict                 6.0.4
opencv-python             4.8.1.78
packaging                 23.2
pefile                    2023.2.7
Pillow                    9.5.0
pip                       23.3.1
pydantic                  2.4.2
pydantic_core             2.10.1
pyinstaller               6.2.0
pyinstaller-hooks-contrib 2023.10
pyparsing                 3.1.1
PyQt5                     5.15.10
PyQt5-Qt5                 5.15.2
PyQt5-sip                 12.13.0
PyQt5-stubs               5.15.6.0
python-dateutil           2.8.2
pywin32-ctypes            0.2.2
replicate                 0.15.5
requests                  2.31.0
rfc3986                   1.5.0
scipy                     1.11.3
setuptools                68.2.0
six                       1.16.0
sniffio                   1.3.0
soupsieve                 2.5
sympy                     1.12
tqdm                      4.66.1
typing_extensions         4.8.0
urllib3                   2.0.7
wheel                     0.41.2
yarl                      1.9.2

requests~=2.31.0
PyQt5~=5.15.10
Pillow~=9.5.0